+++ packages/yourloops/components/chips/patient-diabetic-profile-chip.tsx
@@ -41,7 +41,7 @@ const styles = makeStyles()((theme: Theme) => {
       minWidth: 'auto',
       paddingX: theme.spacing(2),
       paddingY: theme.spacing(0.5),
-      margin: theme.spacing(0, 0, 0, 4),
+      margin: theme.spacing(0, 0, 0, 2),
       height: '26px'
     }
   }
@@ -50,7 +50,7 @@ const styles = makeStyles()((theme: Theme) => {
 
 interface ChipConfig {
   label: string
-  color: 'primary' | 'pink' | 'darkBlue'
+  color: 'lightBlue' | 'lightPink' | 'lightDarkBlue'
 }
 
 interface PatientDiabeticProfileProps {
@@ -68,17 +68,17 @@ export const PatientDiabeticProfileChip: FunctionComponent<PatientDiabeticProfil
       case DiabeticType.DT1DT2:
         return {
           label: t('range-profile-type-1-and-2'),
-          color: 'primary'
+          color: 'lightBlue'
         }
       case DiabeticType.DT1Pregnancy:
         return {
           label: t('range-profile-pregnancy-type-1'),
-          color: 'pink'
+          color: 'lightPink'
         }
       case DiabeticType.Custom:
         return {
           label: t('range-profile-custom'),
-          color: 'darkBlue'
+          color: 'lightDarkBlue'
         }
     }
   }