+++ packages/yourloops/components/patient-list/models/grid-row.model.ts
@@ -28,6 +28,7 @@
 import { type PatientListColumns, type PendingPatientListColumns } from './enums/patient-list.enum'
 import { type Patient } from '../../../lib/patient/models/patient.model'
 import { DiabeticType } from 'medical-domain'
+import moment from 'moment-timezone'
 
 export interface GridRowModel {
   id: string
@@ -43,7 +44,7 @@ export interface GridRowModel {
   [PatientListColumns.GlucoseManagementIndicator]?: number
   [PatientListColumns.BelowRange]?: number
   [PatientListColumns.Variance]?: number
-  [PatientListColumns.LastDataUpdate]?: string
+  [PatientListColumns.LastDataUpdate]?: moment.Moment | null
   [PatientListColumns.Actions]: Patient
   [PatientListColumns.PatientProfile]: DiabeticType
 }