+++ packages/yourloops/test/integration/assert/patient-nav-bar.assert.ts
@@ -50,7 +50,7 @@ export const checkPatientNavBarAsCaregiver = () => {
 }
 
 export const checkPatientNavBarAsPatient = () => {
-  checkPatientNavBarCommon('DashboardDailyTrendsDevice')
+  checkPatientNavBarCommon('DashboardDailyTrendsProfileDevice')
 }
 
 export const checkPatientDropdown = async (initialPatient: Patient, patientToSwitchTo: Patient) => {
@@ -74,7 +74,12 @@ export const checkPatientDropdown = async (initialPatient: Patient, patientToSwi
   expect(secondaryHeader).toHaveTextContent(secondPatientHeaderContent)
 }
 
-export const checkPatientNavBarForPatientAndCaregiver = async () => {
+export const checkPatientNavBarForPatient = async () => {
+  const secondaryHeader = await screen.findByTestId('patient-nav-bar')
+  expect(secondaryHeader).toHaveTextContent('DashboardDailyTrendsProfileDevicesDownload report')
+}
+
+export const checkPatientNavBarForCaregiver = async () => {
   const secondaryHeader = await screen.findByTestId('patient-nav-bar')
   expect(secondaryHeader).toHaveTextContent('DashboardDailyTrendsDevicesDownload report')
 }